+++ CHANGELOG.md
@@ -72,6 +72,7 @@
 * Enclose the query in quotes to allow for parentheses. Fixes UIIN-2516.
 * Make the isShared prop boolean to avoid crashes when calling methods from undefined. Fixes UIIN-2554.
 * Add advanced search query for facets. Fixes UIIN-2536.
+* Show Effective location facet for all Call Number browse sub-options. Fixes UIIN-2499.
 
 ## [9.4.11](https://github.com/folio-org/ui-inventory/tree/v9.4.11) (2023-08-02)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v9.4.10...v9.4.11)